Split simple functions in windows injector to different files
