+++ src/libinjector/win/win_injector.c
@@ -102,454 +102,11 @@
  *                                                                         *
 ***************************************************************************/
 
-#define LIBVMI_EXTRA_GLIB
-#define LIBVMI_EXTRA_JSON
-
-#include <libvmi/libvmi.h>
-#include <libvmi/libvmi_extra.h>
-#include <libvmi/x86.h>
-#include <stdlib.h>
-#include <string.h>
-#include <errno.h>
-#include <sys/mman.h>
-#include <stdio.h>
-#include <signal.h>
-#include <inttypes.h>
-#include <glib.h>
-#include <json-c/json.h>
-
-#include "libdrakvuf/libdrakvuf.h"
-#include <libinjector/libinjector.h>
-#include "private.h"
-#include "win_private.h"
-
-typedef enum
-{
-    INJECT_RESULT_SUCCESS,
-    INJECT_RESULT_TIMEOUT,
-    INJECT_RESULT_CRASH,
-    INJECT_RESULT_PREMATURE,
-    INJECT_RESULT_ERROR_CODE,
-    INJECT_RESULT_INIT_FAIL
-} inject_result_t;
-
-struct injector
-{
-    // Inputs:
-    unicode_string_t* target_file_us;
-    vmi_pid_t target_pid;
-    uint32_t target_tid;
-    unicode_string_t* cwd_us;
-    bool break_loop_on_detection;
-
-    // Internal:
-    drakvuf_t drakvuf;
-    bool is32bit, hijacked, resumed, detected;
-    injection_method_t method;
-    bool global_search;
-    bool wait_for_exit;
-    addr_t exec_func;
-    reg_t target_rsp;
-
-    // For create process
-    addr_t resume_thread;
-
-    // For terminate process
-    vmi_pid_t terminate_pid;
-    addr_t open_process;
-    addr_t exit_process;
-
-    // For shellcode execution
-    addr_t payload, payload_addr, memset;
-    // For readfile/writefile
-    addr_t create_file, read_file, write_file, close_handle, expand_env;
-    size_t binary_size, payload_size;
-    uint32_t status;
-    uint32_t file_handle;
-    FILE* host_file;
-    unicode_string_t* expanded_target;
-
-    // For process doppelganging shellcode
-    addr_t binary, binary_addr, saved_bp;
-    addr_t process_notify;
-
-    const char* binary_path;
-    const char* target_process;
-
-    addr_t process_info;
-    registers_t saved_regs;
-
-    drakvuf_trap_t bp;
-    GSList* memtraps;
-
-    // Used only on x64
-    size_t offsets[OFFSET_MAX];
-
-    // Results:
-    injector_status_t rc;
-    inject_result_t result;
-    struct
-    {
-        bool valid;
-        uint32_t code;
-        const char* string;
-    } error_code;
-
-    uint32_t pid, tid;
-    uint64_t hProc, hThr;
-};
-
-static void free_memtraps(injector_t injector)
-{
-    GSList* loop = injector->memtraps;
-    injector->memtraps = NULL;
-
-    while (loop)
-    {
-        drakvuf_remove_trap(injector->drakvuf, loop->data, (drakvuf_trap_free_t)free);
-        loop = loop->next;
-    }
-    g_slist_free(loop);
-}
-
-static void free_injector(injector_t injector)
-{
-    if (!injector) return;
-
-    PRINT_DEBUG("Injector freed\n");
-
-    free_memtraps(injector);
-
-    vmi_free_unicode_str(injector->target_file_us);
-    vmi_free_unicode_str(injector->cwd_us);
-    vmi_free_unicode_str(injector->expanded_target);
-
-    g_free((void*)injector->binary);
-    g_free((void*)injector->payload);
-    g_free((void*)injector);
-}
-
-void injector_free_win(injector_t injector)
-{
-    free_injector(injector);
-}
-
-#define SW_SHOWNORMAL   1
-#define MEM_COMMIT      0x00001000
-#define MEM_RESERVE     0x00002000
-#define MEM_PHYSICAL    0x00400000
-#define PAGE_EXECUTE_READWRITE  0x40
-#define CREATE_SUSPENDED 0x00000004
-
-struct startup_info_32
-{
-    uint32_t cb;
-    uint32_t lpReserved;
-    uint32_t lpDesktop;
-    uint32_t lpTitle;
-    uint32_t dwX;
-    uint32_t dwY;
-    uint32_t dwXSize;
-    uint32_t dwYSize;
-    uint32_t dwXCountChars;
-    uint32_t dwYCountChars;
-    uint32_t dwFillAttribute;
-    uint32_t dwFlags;
-    uint16_t wShowWindow;
-    uint16_t cbReserved2;
-    uint32_t lpReserved2;
-    uint32_t hStdInput;
-    uint32_t hStdOutput;
-    uint32_t hStdError;
-};
-
-struct startup_info_64
-{
-    uint32_t cb;
-    addr_t lpReserved;
-    addr_t lpDesktop;
-    addr_t lpTitle;
-    uint32_t dwX;
-    uint32_t dwY;
-    uint32_t dwXSize;
-    uint32_t dwYSize;
-    uint32_t dwXCountChars;
-    uint32_t dwYCountChars;
-    uint32_t dwFillAttribute;
-    uint32_t dwFlags;
-    uint16_t wShowWindow;
-    uint16_t cbReserved2;
-    addr_t lpReserved2;
-    addr_t hStdInput;
-    addr_t hStdOutput;
-    addr_t hStdError;
-};
-
-struct process_information_32
-{
-    uint32_t hProcess;
-    uint32_t hThread;
-    uint32_t dwProcessId;
-    uint32_t dwThreadId;
-} __attribute__ ((packed));
-
-struct process_information_64
-{
-    addr_t hProcess;
-    addr_t hThread;
-    uint32_t dwProcessId;
-    uint32_t dwThreadId;
-} __attribute__ ((packed));
-
+#include "win_injector.h"
+#include "win_functions.h"
 #ifdef ENABLE_DOPPELGANGING
-static int patch_payload(injector_t injector, unsigned char* addr)
-{
-    // First byte at which each variable instanciation start in the shellcode.
-    addr_t offset_target_process = 0xa;
-    addr_t offset_binary_buffer = 0xd0b;
-    addr_t offset_binary_size = 0xd22;
-    addr_t tmp_baddr = injector->binary_addr;
-    addr_t tmp_bsize = injector->binary_size;
-    unsigned char* tmp = NULL;
-    unsigned char* patch = NULL;
-    unsigned int size = sizeof(addr_t);
-    unsigned int size_dword = sizeof(uint32_t);
-
-    // Patch targetProcess string (cf. doppelganging.c) by the one provided at the
-    // command line. This variable contains the path of the program that will be
-    // used as a cover. The string is set one byte at a time, every 13 bytes
-    // (instructions' opcodes are between).
-    tmp = (unsigned char*)(addr + offset_target_process);
-    patch = (unsigned char*)injector->target_process;
-
-    while (*patch != '\0')
-    {
-        *tmp = *patch;
-
-        tmp += 13;
-        patch++;
-    }
-    *tmp = '\0';
-
-    // Patch lpBinaryBuffer address (cf. doppelganging.c). It's an address pointing
-    // to the content of the binary file to inject.
-    tmp = (unsigned char*)(addr + offset_binary_buffer);
-
-    for (; size > 0; size--)
-    {
-        *tmp = (unsigned char)tmp_baddr;
-
-        tmp_baddr = tmp_baddr >> 8;
-        tmp++;
-    }
-
-    // Patch binarySize (cf. doppelganging.c). It's an address pointing
-    // to the content of the binary file to inject.
-    tmp = (unsigned char*)(addr + offset_binary_size);
-
-    for (; size_dword > 0; size_dword--)
-    {
-        *tmp = (unsigned char)tmp_bsize;
-
-        tmp_bsize = tmp_bsize >> 8;
-        tmp++;
-    }
-
-    return 0;
-}
+#include "methods/win_dopple.h"
 #endif
-
-static unicode_string_t* convert_utf8_to_utf16(char const* str)
-{
-    if (!str) return NULL;
-
-    unicode_string_t us =
-    {
-        .contents = (uint8_t*)g_strdup(str),
-        .length = strlen(str),
-        .encoding = "UTF-8",
-    };
-
-    if (!us.contents) return NULL;
-
-    unicode_string_t* out = (unicode_string_t*)g_try_malloc0(sizeof(unicode_string_t));
-    if (!out)
-    {
-        g_free(us.contents);
-        return NULL;
-    }
-
-    status_t rc = vmi_convert_str_encoding(&us, out, "UTF-16LE");
-    g_free(us.contents);
-
-    if (VMI_SUCCESS == rc)
-        return out;
-
-    g_free(out);
-    return NULL;
-}
-
-static bool setup_create_process_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[10] = { {0} };
-    struct startup_info_32 si_32 = { 0 };
-    struct process_information_32 pi_32 = { 0 };
-    struct startup_info_64 si_64 = { 0 };
-    struct process_information_64 pi_64 = { 0 };
-
-    // CreateProcess(NULL, TARGETPROC, NULL, NULL, 0, 0, NULL, NULL, &si, pi))
-    init_int_argument(&args[0], 0);
-    init_unicode_argument(&args[1], injector->target_file_us);
-    init_int_argument(&args[2], 0);
-    init_int_argument(&args[3], 0);
-    init_int_argument(&args[4], 0);
-    init_int_argument(&args[5], CREATE_SUSPENDED);
-    init_int_argument(&args[6], 0);
-    init_unicode_argument(&args[7], injector->cwd_us);
-    if (injector->is32bit)
-    {
-        init_struct_argument(&args[8], si_32);
-        init_struct_argument(&args[9], pi_32);
-    }
-    else
-    {
-        init_struct_argument(&args[8], si_64);
-        init_struct_argument(&args[9], pi_64);
-    }
-
-    bool success = setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-    injector->process_info = args[9].data_on_stack;
-    return success;
-}
-
-static bool setup_resume_thread_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[1] = { {0} };
-    init_int_argument(&args[0], injector->hThr);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_shell_execute_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[6] = { {0} };
-
-    // ShellExecute(NULL, NULL, &FilePath, NULL, NULL, SW_SHOWNORMAL)
-    init_int_argument(&args[0], 0);
-    init_unicode_argument(&args[1], NULL);
-    init_unicode_argument(&args[2], injector->target_file_us);
-    init_unicode_argument(&args[3], NULL);
-    init_unicode_argument(&args[4], injector->cwd_us);
-    init_int_argument(&args[5], SW_SHOWNORMAL);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_virtual_alloc_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[4] = { {0} };
-
-    // VirtualAlloc(NULL, size, allocation_type, protect);
-    init_int_argument(&args[0], 0);
-    // Allocate enough space for the shellcode and the binary at once
-    init_int_argument(&args[1], injector->payload_size + injector->binary_size);
-    init_int_argument(&args[2], MEM_COMMIT | MEM_RESERVE);
-    init_int_argument(&args[3], PAGE_EXECUTE_READWRITE);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_memset_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[4] = { {0} };
-
-    // memset(payload_addr, c, payload_size);
-    init_int_argument(&args[0], injector->payload_addr);
-    init_int_argument(&args[1], 0);
-    init_int_argument(&args[2], injector->payload_size + injector->binary_size);
-    init_int_argument(&args[3], 0);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_expand_env_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[3] = { {0} };
-
-    init_unicode_argument(&args[0], injector->target_file_us);
-    init_int_argument(&args[1], injector->payload_addr);
-    init_int_argument(&args[2], 256);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_create_file_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[7] = { {0} };
-
-    init_int_argument(&args[0], injector->payload_addr);
-
-    if (injector->method == INJECT_METHOD_READ_FILE)
-        init_int_argument(&args[1], GENERIC_READ);
-    else
-        init_int_argument(&args[1], GENERIC_WRITE);
-
-    init_int_argument(&args[2], FILE_SHARE_READ);
-    init_int_argument(&args[3], 0);
-
-    if (injector->method == INJECT_METHOD_READ_FILE)
-        init_int_argument(&args[4], OPEN_EXISTING);
-    else
-        init_int_argument(&args[4], CREATE_NEW);
-
-    init_int_argument(&args[5], FILE_ATTRIBUTE_NORMAL);
-    init_int_argument(&args[6], 0);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_read_file_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[5] = { {0} };
-
-    // ReadFile(HANDLE, lpBuffer, nNumberOfBytesToRead, lpNumberOfBytesRead, NULL)
-
-    // first page is reserved for lpNumberOfBytesRead value
-    init_int_argument(&args[0], injector->file_handle);
-    init_int_argument(&args[1], injector->payload_addr + FILE_BUF_RESERVED);
-    init_int_argument(&args[2], injector->payload_size - FILE_BUF_RESERVED);
-    init_int_argument(&args[3], injector->payload_addr);
-    init_int_argument(&args[4], 0);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_write_file_stack(injector_t injector, x86_registers_t* regs, size_t amount)
-{
-    struct argument args[5] = { {0} };
-
-    // WriteFile(HANDLE, lpBuffer, nNumberOfBytesToWrite, lpNumberOfBytesWritten, NULL)
-
-    // first page is reserved for lpNumberOfBytesWritten value
-    init_int_argument(&args[0], injector->file_handle);
-    init_int_argument(&args[1], injector->payload_addr + FILE_BUF_RESERVED);
-    init_int_argument(&args[2], amount);
-    init_int_argument(&args[3], injector->payload_addr);
-    init_int_argument(&args[4], 0);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
-
-static bool setup_close_handle_stack(injector_t injector, x86_registers_t* regs)
-{
-    struct argument args[1] = { {0} };
-
-    // CloseHandle(HANDLE)
-
-    init_int_argument(&args[0], injector->file_handle);
-
-    return setup_stack(injector->drakvuf, regs, args, ARRAY_SIZE(args));
-}
 
 static bool injector_set_hijacked(injector_t injector, drakvuf_trap_info_t* info)
 {
@@ -601,9 +158,6 @@ static void fill_created_process_info(injector_t injector, drakvuf_trap_info_t* 
 
     drakvuf_release_vmi(injector->drakvuf);
 }
-
-static event_response_t injector_int3_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info);
-static event_response_t injector_int3_terminate_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info);
 
 static bool setup_int3_trap(injector_t injector, drakvuf_trap_info_t* info, addr_t bp_addr)
 {
@@ -1047,7 +601,7 @@ static event_response_t inject_payload(drakvuf_t drakvuf, drakvuf_trap_info_t* i
     return 0;
 }
 
-static event_response_t injector_int3_terminate_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
+event_response_t injector_int3_terminate_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
 {
     injector_t injector = info->trap->data;
 
@@ -1192,7 +746,7 @@ static event_response_t injector_int3_terminate_cb(drakvuf_t drakvuf, drakvuf_tr
     return 0;
 }
 
-static event_response_t injector_int3_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
+event_response_t injector_int3_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
 {
     injector_t injector = info->trap->data;
 
@@ -1947,50 +1501,6 @@ static bool inject(drakvuf_t drakvuf, injector_t injector)
     return true;
 }
 
-static bool load_file_to_memory(addr_t* output, size_t* size, const char* file)
-{
-    if (!output || !size || !file)
-        return false;
-
-    long payload_size = 0;
-    unsigned char* data = NULL;
-    FILE* fp = fopen(file, "rb");
-
-    if (!fp)
-        return false;
-
-    // obtain file size:
-    fseek (fp, 0, SEEK_END);
-    if ( (payload_size = ftell (fp)) < 0 )
-    {
-        fclose(fp);
-        return false;
-    }
-    rewind (fp);
-
-    data = g_try_malloc0(payload_size);
-    if ( !data )
-    {
-        fclose(fp);
-        return false;
-    }
-
-    if ( (size_t)payload_size != fread(data, payload_size, 1, fp) )
-    {
-        g_free(data);
-        fclose(fp);
-        return false;
-    }
-
-    *output = (addr_t)data;
-    *size = payload_size;
-
-    PRINT_DEBUG("Size of file read: %lu\n", payload_size);
-
-    fclose(fp);
-
-    return true;
-}
 
 static void print_injection_info(output_format_t format, const char* file, injector_t injector)
 {
@@ -2220,64 +1730,6 @@ static void print_injection_info(output_format_t format, const char* file, injec
     g_free(escaped_pname);
     g_free(escaped_arguments);
     g_strfreev(split_results);
-}
-
-struct module_context
-{
-    const char* lib;
-    const char* fun;
-    addr_t module_addr;
-    addr_t addr;
-};
-
-static bool module_visitor(drakvuf_t drakvuf, const module_info_t* module_info, void* ctx )
-{
-    struct module_context* data = (struct module_context*)ctx;
-
-    if (module_info->base_addr != data->module_addr)
-        return false;
-
-    data->addr = drakvuf_exportsym_to_va(drakvuf, module_info->eprocess_addr, data->lib, data->fun);
-    if (data->addr)
-        return true;
-
-    return false;
-}
-
-static addr_t get_function_va(drakvuf_t drakvuf, addr_t eprocess_base, char const* lib, char const* fun, bool global_search)
-{
-    // First check current process for function
-    addr_t addr = drakvuf_exportsym_to_va(drakvuf, eprocess_base, lib, fun);
-    if (addr)
-        return addr;
-
-    // If function is not mapped into the processes address space search it in other processes
-    struct module_context module_ctx =
-    {
-        .lib = lib,
-        .fun = fun,
-        .addr = 0
-    };
-
-    if (global_search)
-    {
-        // First get modules load address to search for other process with same address
-        ACCESS_CONTEXT(ctx);
-        ctx.translate_mechanism = VMI_TM_PROCESS_PID;
-
-        addr_t module_list_head;
-        if (drakvuf_get_process_pid(drakvuf, eprocess_base, &ctx.pid) &&
-            drakvuf_get_module_list(drakvuf, eprocess_base, &module_list_head) &&
-            drakvuf_get_module_base_addr_ctx(drakvuf, module_list_head, &ctx, lib, &module_ctx.module_addr))
-        {
-            drakvuf_enumerate_processes_with_module(drakvuf, lib, module_visitor, &module_ctx);
-        }
-    }
-
-    if (!module_ctx.addr)
-        PRINT_DEBUG("Failed to get address of %s!%s\n", lib, fun);
-
-    return module_ctx.addr;
 }
 
 static bool initialize_injector_functions(drakvuf_t drakvuf, injector_t injector, const char* file, const char* binary_path)
