+++ src/libinjector/win/win_private.h
