+++ src/libinjector/win/win_utils.h
@@ -1,0 +1,290 @@
+/*********************IMPORTANT DRAKVUF LICENSE TERMS**********************
+ *                                                                         *
+ * DRAKVUF (C) 2014-2021 Tamas K Lengyel.                                  *
+ * Tamas K Lengyel is hereinafter referred to as the author.               *
+ * This program is free software; you may redistribute and/or modify it    *
+ * under the terms of the GNU General Public License as published by the   *
+ * Free Software Foundation; Version 2 ("GPL"), BUT ONLY WITH ALL OF THE   *
+ * CLARIFICATIONS AND EXCEPTIONS DESCRIBED HEREIN.  This guarantees your   *
+ * right to use, modify, and redistribute this software under certain      *
+ * conditions.  If you wish to embed DRAKVUF technology into proprietary   *
+ * software, alternative licenses can be acquired from the author.         *
+ *                                                                         *
+ * Note that the GPL places important restrictions on "derivative works",  *
+ * yet it does not provide a detailed definition of that term.  To avoid   *
+ * misunderstandings, we interpret that term as broadly as copyright law   *
+ * allows.  For example, we consider an application to constitute a        *
+ * derivative work for the purpose of this license if it does any of the   *
+ * following with any software or content covered by this license          *
+ * ("Covered Software"):                                                   *
+ *                                                                         *
+ * o Integrates source code from Covered Software.                         *
+ *                                                                         *
+ * o Reads or includes copyrighted data files.                             *
+ *                                                                         *
+ * o Is designed specifically to execute Covered Software and parse the    *
+ * results (as opposed to typical shell or execution-menu apps, which will *
+ * execute anything you tell them to).                                     *
+ *                                                                         *
+ * o Includes Covered Software in a proprietary executable installer.  The *
+ * installers produced by InstallShield are an example of this.  Including *
+ * DRAKVUF with other software in compressed or archival form does not     *
+ * trigger this provision, provided appropriate open source decompression  *
+ * or de-archiving software is widely available for no charge.  For the    *
+ * purposes of this license, an installer is considered to include Covered *
+ * Software even if it actually retrieves a copy of Covered Software from  *
+ * another source during runtime (such as by downloading it from the       *
+ * Internet).                                                              *
+ *                                                                         *
+ * o Links (statically or dynamically) to a library which does any of the  *
+ * above.                                                                  *
+ *                                                                         *
+ * o Executes a helper program, module, or script to do any of the above.  *
+ *                                                                         *
+ * This list is not exclusive, but is meant to clarify our interpretation  *
+ * of derived works with some common examples.  Other people may interpret *
+ * the plain GPL differently, so we consider this a special exception to   *
+ * the GPL that we apply to Covered Software.  Works which meet any of     *
+ * these conditions must conform to all of the terms of this license,      *
+ * particularly including the GPL Section 3 requirements of providing      *
+ * source code and allowing free redistribution of the work as a whole.    *
+ *                                                                         *
+ * Any redistribution of Covered Software, including any derived works,    *
+ * must obey and carry forward all of the terms of this license, including *
+ * obeying all GPL rules and restrictions.  For example, source code of    *
+ * the whole work must be provided and free redistribution must be         *
+ * allowed.  All GPL references to "this License", are to be treated as    *
+ * including the terms and conditions of this license text as well.        *
+ *                                                                         *
+ * Because this license imposes special exceptions to the GPL, Covered     *
+ * Work may not be combined (even as part of a larger work) with plain GPL *
+ * software.  The terms, conditions, and exceptions of this license must   *
+ * be included as well.  This license is incompatible with some other open *
+ * source licenses as well.  In some cases we can relicense portions of    *
+ * DRAKVUF or grant special permissions to use it in other open source     *
+ * software.  Please contact tamas.k.lengyel@gmail.com with any such       *
+ * requests.  Similarly, we don't incorporate incompatible open source     *
+ * software into Covered Software without special permission from the      *
+ * copyright holders.                                                      *
+ *                                                                         *
+ * If you have any questions about the licensing restrictions on using     *
+ * DRAKVUF in other works, are happy to help.  As mentioned above,         *
+ * alternative license can be requested from the author to integrate       *
+ * DRAKVUF into proprietary applications and appliances.  Please email     *
+ * tamas.k.lengyel@gmail.com for further information.                      *
+ *                                                                         *
+ * If you have received a written license agreement or contract for        *
+ * Covered Software stating terms other than these, you may choose to use  *
+ * and redistribute Covered Software under those terms instead of these.   *
+ *                                                                         *
+ * Source is provided to this software because we believe users have a     *
+ * right to know exactly what a program is going to do before they run it. *
+ * This also allows you to audit the software for security holes.          *
+ *                                                                         *
+ * Source code also allows you to port DRAKVUF to new platforms, fix bugs, *
+ * and add new features.  You are highly encouraged to submit your changes *
+ * on https://github.com/tklengyel/drakvuf, or by other methods.           *
+ * By sending these changes, it is understood (unless you specify          *
+ * otherwise) that you are offering unlimited, non-exclusive right to      *
+ * reuse, modify, and relicense the code.  DRAKVUF will always be          *
+ * available Open Source, but this is important because the inability to   *
+ * relicense code has caused devastating problems for other Free Software  *
+ * projects (such as KDE and NASM).                                        *
+ * To specify special license conditions of your contributions, just say   *
+ * so when you send them.                                                  *
+ *                                                                         *
+ * This program is distributed in the hope that it will be useful, but     *
+ * WITHOUT ANY WARRANTY; without even the implied warranty of              *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the DRAKVUF   *
+ * license file for more details (it's in a COPYING file included with     *
+ * DRAKVUF, and also available from                                        *
+ * https://github.com/tklengyel/drakvuf/COPYING)                           *
+ *                                                                         *
+***************************************************************************/
+
+#ifndef WIN_UTILS_H
+#define WIN_UTILS_H
+
+#define LIBVMI_EXTRA_GLIB
+#define LIBVMI_EXTRA_JSON
+
+#include <libvmi/libvmi.h>
+#include <libvmi/libvmi_extra.h>
+#include <libvmi/x86.h>
+#include <stdlib.h>
+#include <string.h>
+#include <errno.h>
+#include <sys/mman.h>
+#include <stdio.h>
+#include <signal.h>
+#include <inttypes.h>
+#include <glib.h>
+#include <json-c/json.h>
+
+#include "libdrakvuf/libdrakvuf.h"
+#include <libinjector/libinjector.h>
+#include "private.h"
+#include "win_private.h"
+
+#define SW_SHOWNORMAL   1
+#define MEM_COMMIT      0x00001000
+#define MEM_RESERVE     0x00002000
+#define MEM_PHYSICAL    0x00400000
+#define PAGE_EXECUTE_READWRITE  0x40
+#define CREATE_SUSPENDED 0x00000004
+
+typedef enum
+{
+    INJECT_RESULT_SUCCESS,
+    INJECT_RESULT_TIMEOUT,
+    INJECT_RESULT_CRASH,
+    INJECT_RESULT_PREMATURE,
+    INJECT_RESULT_ERROR_CODE,
+    INJECT_RESULT_INIT_FAIL
+} inject_result_t;
+
+struct injector
+{
+    // Inputs:
+    unicode_string_t* target_file_us;
+    vmi_pid_t target_pid;
+    uint32_t target_tid;
+    unicode_string_t* cwd_us;
+    bool break_loop_on_detection;
+
+    // Internal:
+    drakvuf_t drakvuf;
+    bool is32bit, hijacked, resumed, detected;
+    injection_method_t method;
+    bool global_search;
+    bool wait_for_exit;
+    addr_t exec_func;
+    reg_t target_rsp;
+
+    // For create process
+    addr_t resume_thread;
+
+    // For terminate process
+    vmi_pid_t terminate_pid;
+    addr_t open_process;
+    addr_t exit_process;
+
+    // For shellcode execution
+    addr_t payload, payload_addr, memset;
+    // For readfile/writefile
+    addr_t create_file, read_file, write_file, close_handle, expand_env;
+    size_t binary_size, payload_size;
+    uint32_t status;
+    uint32_t file_handle;
+    FILE* host_file;
+    unicode_string_t* expanded_target;
+
+    // For process doppelganging shellcode
+    addr_t binary, binary_addr, saved_bp;
+    addr_t process_notify;
+
+    const char* binary_path;
+    const char* target_process;
+
+    addr_t process_info;
+    registers_t saved_regs;
+
+    drakvuf_trap_t bp;
+    GSList* memtraps;
+
+    // Used only on x64
+    size_t offsets[OFFSET_MAX];
+
+    // Results:
+    injector_status_t rc;
+    inject_result_t result;
+    struct
+    {
+        bool valid;
+        uint32_t code;
+        const char* string;
+    } error_code;
+
+    uint32_t pid, tid;
+    uint64_t hProc, hThr;
+};
+
+
+struct startup_info_32
+{
+    uint32_t cb;
+    uint32_t lpReserved;
+    uint32_t lpDesktop;
+    uint32_t lpTitle;
+    uint32_t dwX;
+    uint32_t dwY;
+    uint32_t dwXSize;
+    uint32_t dwYSize;
+    uint32_t dwXCountChars;
+    uint32_t dwYCountChars;
+    uint32_t dwFillAttribute;
+    uint32_t dwFlags;
+    uint16_t wShowWindow;
+    uint16_t cbReserved2;
+    uint32_t lpReserved2;
+    uint32_t hStdInput;
+    uint32_t hStdOutput;
+    uint32_t hStdError;
+};
+
+struct startup_info_64
+{
+    uint32_t cb;
+    addr_t lpReserved;
+    addr_t lpDesktop;
+    addr_t lpTitle;
+    uint32_t dwX;
+    uint32_t dwY;
+    uint32_t dwXSize;
+    uint32_t dwYSize;
+    uint32_t dwXCountChars;
+    uint32_t dwYCountChars;
+    uint32_t dwFillAttribute;
+    uint32_t dwFlags;
+    uint16_t wShowWindow;
+    uint16_t cbReserved2;
+    addr_t lpReserved2;
+    addr_t hStdInput;
+    addr_t hStdOutput;
+    addr_t hStdError;
+};
+
+struct process_information_32
+{
+    uint32_t hProcess;
+    uint32_t hThread;
+    uint32_t dwProcessId;
+    uint32_t dwThreadId;
+} __attribute__ ((packed));
+
+struct process_information_64
+{
+    addr_t hProcess;
+    addr_t hThread;
+    uint32_t dwProcessId;
+    uint32_t dwThreadId;
+} __attribute__ ((packed));
+
+unicode_string_t* convert_utf8_to_utf16(char const* str);
+bool load_file_to_memory(addr_t* output, size_t* size, const char* file);
+addr_t get_function_va(drakvuf_t drakvuf, addr_t eprocess_base, char const* lib, char const* fun, bool global_search);
+void free_memtraps(injector_t injector);
+void free_injector(injector_t injector);
+void injector_free_win(injector_t injector);
+
+struct module_context
+{
+    const char* lib;
+    const char* fun;
+    addr_t module_addr;
+    addr_t addr;
+};
+
+bool module_visitor(drakvuf_t drakvuf, const module_info_t* module_info, void* ctx );
+
+#endif
